SI-6775 Racy synchronization of specialized methods.
When the new specialized method is created that overrides a method
in a base class, cloneSymbol is called on a base method, rather than
on a non-specialized overrider.

This results in flags incorrectly (un)set on the clone, which include,
among others, SYNCHRONIZED snd FINAL.

To fix the problem, we don't forward to specializedOverload, but
rather clone overriding method directly.