Add support for java8 default methods.
In classfile parser: mark symbols which represent interface
methods yet have code attributes with new flag DEFAULTMETHOD.
These must be kept distinct from regular method bodies so that
an error can be issued when a regular concrete method is
overridden without the override keyword, but not when the
overridden method is a default.

In java source parser: mark Modifiers of interface default
methods with DEFAULTMETHOD flag.