temporarily breaks SI-5353
The subsequent fix to SI-5923 unmasks the fact that SI-5353 has not
been fixed - it's just that one of its manifestation got hidden behing
SI-5923. In fact, if in the code snippet from the bug report we change
Array() to Array[Nothing](), things will start crashing as usual.

Therefore in this pull request, which is dedicated to SI-5923,
I have to disable files/neg/t5353.scala, leaving its fix to a
separate patch. See a discussion about a potential fix here:
https://github.com/scala/scala/pull/2466.