SI-5634 eliminate overly verbose error message
In fact I couldn't see where the error message described in
the ticket is being produced, and I'm supposing it has
already vacated the premises. I went through ClassfileParser
and UnPickler, the two enclaves most likely to harbor the
spewer, and hardened the output. No printlns, tighter
assertion messages, and string interpolation for aesthetics.
We're not savages.