use relative symlink in distpack
To simplify building a release on jenkins, we run distpack-opt in one job,
store the `dists/` directory in a tar ball, archive that artifact and
copy it to the downstream jobs that package on windows and unix.

To make the tarball portable between machines, it must not use absolute symlinks.