SI-7445 element order of ListMap.iterator should be reversed
Since the underlying collection of ListMap is indeed a List,
order-related methods of ListMap should respect the order
of elements in the underlying List.

Since the underlying List of ListMap stores elements in a
reversed input order (i.e. LIFO), it is natural that
ListMap.iterator also returns elements in the same order.