SI-7532 Fix regression in Java inner classfile reader
395e90a modified the detection of top-level classes in
ClassfileParser in two ways:
  1. used `Name#containsChar` rather than `toString.indexOf ...` (good!)
  2. decoded the name before doing this check (bad!)

That code is actually only run for non-Scala classfiles, whose
names don't need decoding. Attempting to do so converted `R$attr`
to `R@tr`, which no longer contains a '$', and was wrongly treated
as a top level class.

This commit reverts the use of `decodedName`, and inlines the method
to its only call site for clarity.