SI-7517 type constructors too eagerly normalized.
I think 403eadd0f1 was largely a symptomatic remedy
(not that we shouldn't harden against such outcomes)
and that this commit gets closer to the root causes.
The unanticipated change to test/files/run/t6113.check
is like a cry of support from the jury box.

  -Foo[[X](Int, X)]
  +Foo[AnyRef{type l[X] = (Int, X)}#l]

We should continue to look at calls to normalize with
grave suspicion.